yomama/requirements.txt
api==0.0.7
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-import-export==3.0.2
django-ninja==0.20.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mysqlclient==2.1.1
nose==1.3.7
odfpy==1.4.1
openpyxl==3.1.1
orjson==3.8.6
pydantic==1.10.5
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
typing_extensions==4.5.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
